feat(desktop): discover category independent page
